requirement.txt
langchain
python-dotenv
ipykernel
langchain_community
pypdf
bs4
arxiv
PyMuPDF
wikipedia
langchain-text-splitters
langchain-openai
chromadb
langchain-community
langchain-ollama
sentence_transformers
langchain_huggingface
tf-keras
faiss-cpu
langchain_chroma
beautifulsoup4
streamlit
langchain_groq
langchain_core
fastapi
uvicorn
langserve
sse_starlette
duckduckgo-search
huggingface_hub